Booting from a clean elpa folder.
There were some assumptions that packages were already installed in the
elpa folder. This were mainly toplevel require's which are executed when
loading the .el files.

Additionally there were some calls to functions which are now wrapped
with a call to eval-after-load of the respective library.

The package loading was considerably revamped again, maybe unnecessary
so. In any case the code was still there uncommented. Cleaned that up so
it compiles cleanly.