Add elasticsearch support
Added support for elasticsearch mode.

This commit also contains an advice function to login using basic
authentication with the normal url-retrieve method for packages
which do not support this and without the need to patch the code.

This was needed for elasticsearch to log into hunter.melexis.com but
can be used for other applications too, in principle.